Conflicts:
	src/Pi.cpp
	src/Sector.cpp

+++ src/Sector.cpp
@@ -173,7 +173,7 @@ void Sector::GenerateSystems() const
 				s.seed = 0;
 				s.customSystem = 0;
 				
-				float spec = rng.Int32(1000000);
+				double spec = rng.Double(1000000.0);
 				// frequencies from wikipedia
 				/*if (spec < 100) { // should be 1 but that is boring
 					s.starType[0] = SBody::TYPE_STAR_O;
@@ -196,7 +196,7 @@ void Sector::GenerateSystems() const
 				}*/
 				//if ((sx > 50) || (sx < -50) ||
 				//	(sy > 50) || (sy < -50))
-	
+
 				// Frequencies are low enough that we probably don't need this anymore.
 				if (isqrt(1+sx*sx+sy*sy) > 10) 
 				{
@@ -371,7 +371,7 @@ void Sector::GenerateSystems() const
 float Sector::DistanceBetween(const Sector *a, int sysIdxA, const Sector *b, int sysIdxB)
 {
 	vector3f dv = a->GetSystem(sysIdxA).p - b->GetSystem(sysIdxB).p;
-	dv += Sector::SIZE*vector3f(a->sx - b->sx, a->sy - b->sy, a->sz - b->sz);
+	dv += Sector::SIZE*vector3f(float(a->sx - b->sx), float(a->sy - b->sy), float(a->sz - b->sz));
 	return dv.Length();
 }
 
@@ -390,9 +390,9 @@ std::string Sector::GenName(System &sys, MTRand &rng) const
 		case SBody::TYPE_STAR_K: chance += 8*dist; break;
 		case SBody::TYPE_STAR_O_GIANT:
 		case SBody::TYPE_STAR_B_GIANT: chance = 50; break;
-		case SBody::TYPE_STAR_A_GIANT: chance = 0.2*dist; break;
-		case SBody::TYPE_STAR_F_GIANT: chance = 0.4*dist; break;
-		case SBody::TYPE_STAR_G_GIANT: chance = 0.5*dist; break;
+		case SBody::TYPE_STAR_A_GIANT: chance = int(0.2*dist); break;
+		case SBody::TYPE_STAR_F_GIANT: chance = int(0.4*dist); break;
+		case SBody::TYPE_STAR_G_GIANT: chance = int(0.5*dist); break;
 		case SBody::TYPE_STAR_K_GIANT:
 		case SBody::TYPE_STAR_M_GIANT: chance = dist; break;
 		case SBody::TYPE_STAR_O_SUPER_GIANT:
